Move observations in to their own component
