+++ app/components/observation_component/observation_base.html.erb
