+++ app/components/timeline_component/timeline_component.html.erb
@@ -16,7 +16,7 @@
           <tr scope="row">
             <td></td>
             <td class="timeline-border-left"></td>
-            <%= component 'timeline/observation', observation: observation, show_actions: show_actions %>
+            <%= component 'observation', observation: observation, show_actions: show_actions %>
           </tr>
         <% end %>
       </tbody>
